// src/core/configs/enums.ts
export enum TypeOrmEnvKeys {
  Connection = 'TYPEORM_CONNECTION',
  Host = 'TYPEORM_HOST',
  Port = 'TYPEORM_PORT',
  Username = 'TYPEORM_USERNAME',
  Password = 'TYPEORM_PASSWORD',
  Database = 'TYPEORM_DATABASE',
  Synchorinize = 'TYPEORM_SYNCHRONIZE',
  Logging = 'TYPEORM_LOGGING',
  Entities = 'TYPEORM_ENTITIES',
  Migrations = 'TYPEORM_MIGRATIONS',
  Timezone = 'TYPEORM_TIMEZONE',
}